pyproject.toml
[project]
name = "discord-python-starter"
version = "0.1.0"
description = "Discord Python Starter Template"
authors = ["Vineeth Voruganti <13438633+VVoruganti@users.noreply.github.com>", "Ben <ben@plasticlabs.ai>"]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "py-cord>=2.5.0",
    "python-dotenv>=1.0.1",
    "openai>=1.42.0",
    "schedule>=1.2.2",
    "requests>=2.32.3",
    "honcho-ai>=1.2.1",
]
